fast_api/requirements.txt
fastapi==0.100.0
pydantic==1.10.2
python-dotenv==1.0.0
uvicorn==0.22.0
google-generativeai==0.1.0
shutil
re
pathlib
json